+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [2.1.0-dev.5](https://github.com/point-source/flywheel/compare/v2.1.0-dev.4...v2.1.0-dev.5) (2026-06-18)
+
+
+### Features
+
+* three-state doctor repo-field reads ([#255](https://github.com/point-source/flywheel/issues/255)) ([459ffeb](https://github.com/point-source/flywheel/commit/459ffeb8bcaf8927348cbd3f217635c4eb115eda)), closes [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239) [#239](https://github.com/point-source/flywheel/issues/239)
+
+# [2.1.0-dev.4](https://github.com/point-source/flywheel/compare/v2.1.0-dev.3...v2.1.0-dev.4) (2026-06-18)
+
+
+### Features
+
+* break Dependabot PR deadlock on required conventional-commit check ([#254](https://github.com/point-source/flywheel/issues/254)) ([58f9cfc](https://github.com/point-source/flywheel/commit/58f9cfc505a3b8540340077bf68231c00ca34f2a)), closes [#243](https://github.com/point-source/flywheel/issues/243) [#243](https://github.com/point-source/flywheel/issues/243)
+
 # [2.1.0-dev.3](https://github.com/point-source/flywheel/compare/v2.1.0-dev.2...v2.1.0-dev.3) (2026-06-13)